ENH-005: Reconfigured the layers
Defined the layers for the intial use defined.
Added some sample code which will be included in the final version of the code.
Updated the README file with the details for the next release.
Updated the README with preference for the release
Added some more sample code for implementation of the keyboard.
Create a new keyboard configuration using Colemak-modDH as the default layer.
Added the QWERTY layer for backup.
Added a Symbols layer, Number layer and Function & Bluetooth layer.
Used various mod-morph behavior.
Added various configurations from the following repos:
  * https://github.com/urob/zmk-config
  * https://github.com/minusfive/zmk-config
Removed the homerow mods for the Bluetooth layer on the left split keyboard
Added the Bluetooth profile selection defines for use in homerow mods
Updated the mod-morph reference for Grave/Escape key sequence
Updated the Homerow mods for BT layer to only work as control keys.
Moved the BT profile selection to the next left lower keys
Reconfigured the Colemak and Qwerty layers
Added the custom Escape-Grave key
Updated the escgrave mod-morph to send the command key while using the mod